+++ modules/Food_Service/Accounts.php
@@ -1,5 +1,4 @@
 <?php
-require_once 'modules/Food_Service/includes/DeletePromptX.fnc.php';
 
 if ( ! $_SESSION['FSA_type'])
 	$_SESSION['FSA_type'] = 'student';