+++ modules/Food_Service/ActivityReport.php
@@ -35,31 +35,25 @@ User('PROFILE')=='student'?'':DrawHeader($header);
 
 if ( $_REQUEST['modfunc']=='delete' && AllowEdit())
 {
-	require_once 'modules/Food_Service/includes/DeletePromptX.fnc.php';
 	if ( $_REQUEST['item_id']!='')
 	{
-//FJ add translation
-		if (DeletePromptX(_('Transaction Item')))
+		if ( DeletePrompt( _( 'Transaction Item' ) ) )
 		{
 			require_once 'modules/Food_Service/includes/DeleteTransactionItem.fnc.php';
 			DeleteTransactionItem($_REQUEST['transaction_id'],$_REQUEST['item_id'],$_REQUEST['type']);
 			DBQuery('BEGIN; '.$sql1.'; '.$sql2.'; '.$sql3.'; COMMIT');
 			unset($_REQUEST['modfunc']);
-			unset($_REQUEST['delete_ok']);
 			unset($_SESSION['_REQUEST_vars']['modfunc']);
-			unset($_SESSION['_REQUEST_vars']['delete_ok']);
 		}
 	}
 	else
 	{
-		if (DeletePromptX(_('Transaction')))
+		if ( DeletePrompt( _( 'Transaction' ) ) )
 		{
 			require_once 'modules/Food_Service/includes/DeleteTransaction.fnc.php';
 			DeleteTransaction($_REQUEST['transaction_id'],$_REQUEST['type']);
 			unset($_REQUEST['modfunc']);
-			unset($_REQUEST['delete_ok']);
 			unset($_SESSION['_REQUEST_vars']['modfunc']);
-			unset($_SESSION['_REQUEST_vars']['delete_ok']);
 		}
 	}
 }