+++ modules/Food_Service/Users/ServeMenus.php
@@ -10,8 +10,12 @@ if ( $_REQUEST['modfunc']=='submit')
 {
 	if ( $_REQUEST['submit']['cancel'])
 	{
-		if (DeletePromptX(_('Sale'),_('Cancel')))
-			unset($_SESSION['FSA_sale']);
+		if ( DeletePrompt( _( 'Sale' ), _( 'Cancel' ) ) )
+		{
+			unset( $_SESSION['FSA_sale'] );
+
+			$_REQUEST['modfunc'] = false;
+		}
 	}
 	elseif ( $_REQUEST['submit']['save'])
 	{