+++ modules/Grades/EditHistoryMarkingPeriods.php
@@ -1,6 +1,5 @@
 <?php
 
-include 'modules/Grades/DeletePromptX.fnc.php';
 DrawHeader(ProgramTitle());
 
 if ( $_REQUEST['modfunc']=='update')
@@ -58,11 +57,15 @@ if ( $_REQUEST['modfunc']=='update')
 
 if ( $_REQUEST['modfunc']=='remove')
 {
-	//FJ add translation
-	if (DeletePromptX(_('History Marking Period')))
+	if ( DeletePrompt( _( 'History Marking Period' ) ) )
 	{
-		DBQuery("DELETE FROM history_marking_periods WHERE MARKING_PERIOD_ID='".$_REQUEST['id']."'");
-		DBQuery("DELETE FROM student_report_card_grades WHERE MARKING_PERIOD_ID='".$_REQUEST['id']."'");
+		DBQuery( "DELETE FROM history_marking_periods
+			WHERE MARKING_PERIOD_ID='" . $_REQUEST['id'] . "'" );
+
+		DBQuery( "DELETE FROM student_report_card_grades
+			WHERE MARKING_PERIOD_ID='" . $_REQUEST['id'] . "'" );
+
+		$_REQUEST['modfunc'] = false;
 	}
 }  
 
@@ -125,35 +128,57 @@ function makeTextInput($value,$name)
 //        $extra = 'size=5 maxlength=5';
        
 //    else
-    if ( $name=='NAME')
-	$extra = 'size=20 maxlength=25';
-    else
-    	$extra = 'size=10 maxlength=10';
+	if ( $name=='NAME')
+	{
+		$extra = 'size=20 maxlength=25';
+	}
+	else
+		$extra = 'size=10 maxlength=10';
 
-    return TextInput($value,"values[ $id ][ $name ]",'',$extra);
+	return TextInput(
+		$value,
+		'values[' . $id . '][' . $name . ']',
+		'',
+		$extra
+	);
 }
 
 function makeDateInput($value,$name)
 {    global $THIS_RET;
 
-    if ( $THIS_RET['MARKING_PERIOD_ID'])
-        $id = $THIS_RET['MARKING_PERIOD_ID'];
-    else
-        $id = 'new';
-    return DateInput($value,"values[ $id ][ $name ]",'');
+	if ( $THIS_RET['MARKING_PERIOD_ID'])
+	{
+		$id = $THIS_RET['MARKING_PERIOD_ID'];
+	}
+	else
+		$id = 'new';
+
+	return DateInput(
+		$value,
+		'values[' . $id . '][' . $name . ']',
+		''
+    );
 }
 
 function makeSelectInput($value,$name)
 {    global $THIS_RET;
 
     if ( $THIS_RET['MARKING_PERIOD_ID'])
+    {
         $id = $THIS_RET['MARKING_PERIOD_ID'];
+	}
     else
-        $id = 'new';
+		$id = 'new';
 
-    $options = array('year' => _('Year'), 'semester' => _('Semester'), 'quarter' => _('Quarter'));
+	$options = array('year' => _('Year'), 'semester' => _('Semester'), 'quarter' => _('Quarter'));
 
-    return SelectInput(trim($value),"values[ $id ][ $name ]",'',$options,false);
+	return SelectInput(
+		trim( $value ),
+		'values[' . $id . '][' . $name . ']',
+		'',
+		$options,
+		false
+	);
 }
 
 function makeSchoolYearSelectInput( $value, $name )
@@ -173,5 +198,11 @@ function makeSchoolYearSelectInput( $value, $name )
 		//FJ school year over one/two calendar years format
 		$options[ $year ] = FormatSyear( $year, Config( 'SCHOOL_SYEAR_OVER_2_YEARS' ) );
 
-	return SelectInput( trim( $value ), "values[" . $id . "][" . $name . "]", '', $options, false );
+	return SelectInput(
+		trim( $value ),
+		'values[' . $id . '][' . $name . ']',
+		'',
+		$options,
+		false
+	);
 }