+++ CHANGES.md
@@ -52,6 +52,7 @@ Changes in 2.9
 - Highlight color: add CSS to HTML head in Warehouse.php
 - Create school_fields_seq Sequence in rosariosis.sql & Update.fnc.php
 - Add Fields (and Field Categories) functions in ProgramFunctions/Fields.fnc.php
+- Use DeletePrompt() & Prompt() instead of DeletePromptX() & PromptX() program wide
 
 Changes in 2.8.11
 -----------------